+++ setup.cfg
@@ -11,7 +11,6 @@ packages = find:
 python_requires = >=3.6.0
 install_requires =
     arrayfire>=3.8rc1
-    daal4py==2021.3.0
     numpy>=1.18.2
     pyyaml>=5.4.1
     scipy
