+++ tests/test_gaussian_random_1.py
@@ -1,55 +1,39 @@
 import sklearn
 import sklearn.preprocessing
 import sklearn.random_projection
-import daal4py.sklearn
 import time
 import afsklearn
 import numpy as np
 import pickle
 from afsklearn.patcher import Patcher
 
+def bench_gaussian_random_projection(nbench=1):
 
-nbench = 1
-#rng = np.random.RandomState(0)
-#X = np.sort(rng.normal(loc=0.5, scale=0.25, size=(548, 100000)), axis=0)
+    with open('training_data.pickle', 'rb') as f:
+        X = pickle.load(f)
 
-with open('training_data.pickle', 'rb') as f:
-    X = pickle.load(f)
-
-
-#tic = time.perf_counter()
-#import pdb; pdb.set_trace()
-#for n in range(nbench):
-    #transformer = sklearn.random_projection.GaussianRandomProjection()
-    #X_new = transformer.fit_transform(X)
-    #X_new.shape
-#toc = time.perf_counter()
-#print(f"sklearn fit time {(toc - tic)/nbench:0.4f} seconds")
-
-#daal4py.sklearn.patch_sklearn()
-#tic = time.perf_counter()
-#for n in range(nbench):
-#    qt = sklearn.preprocessing.RobustScaler()
-#    qt.fit_transform(X)
-#toc = time.perf_counter()
-#print(f"daal4py sklearn fit time {(toc - tic)/nbench:0.4f} seconds")
-#
-#daal4py.sklearn.unpatch_sklearn()
-#
-
-Patcher.patch('gaussian_random_projection')
-tic = time.perf_counter()
-for n in range(nbench):
-    transformer = sklearn.random_projection.GaussianRandomProjection(eps=0.1, random_state=0)
-    X_new = transformer.fit_transform(X)
-    y = np.random.randint(2, size=X_new.shape[0])
-
-    from sklearn.naive_bayes import GaussianNB
     tic = time.perf_counter()
-    clf = GaussianNB()
-    clf.fit(X_new,y)
+    for n in range(nbench):
+        transformer = sklearn.random_projection.GaussianRandomProjection()
+        X_new = transformer.fit_transform(X)
+        print(X_new.shape)
     toc = time.perf_counter()
-print(f"afsklearn fit time {(toc - tic)/nbench:0.4f} seconds")
-Patcher.rollback('gaussian_random_projection')
+    sklearn_time = toc - tic
+    print(f"sklearn fit time {(sklearn_time)/nbench:0.4f} seconds")
 
-#print(np.max(np.abs(res - res_af)))
+
+    Patcher.patch('gaussian_random_projection')
+    tic = time.perf_counter()
+    for n in range(nbench):
+        transformer = sklearn.random_projection.GaussianRandomProjection(eps=0.1, random_state=0)
+        X_new_af = transformer.fit_transform(X)
+        print(X_new_af.shape)
+    toc = time.perf_counter()
+    afsklearn_time = toc - tic
+    print(f"afsklearn fit time {(afsklearn_time)/nbench:0.4f} seconds")
+    Patcher.rollback('gaussian_random_projection')
+
+    print(np.max(np.abs(X_new - X_new_af)))
+    return sklearn_time/nbench, afsklearn_time/nbench
+
+def
