+++ tests/test_quantile_1.py
@@ -1,5 +1,4 @@
 import sklearn
-import daal4py.sklearn
 import time
 import afsklearn
 import numpy as np
@@ -15,16 +14,6 @@ for n in range(nbench):
     res = qt.fit_transform(X)
 toc = time.perf_counter()
 print(f"sklearn fit time {(toc - tic)/nbench:0.4f} seconds")
-
-#daal4py.sklearn.patch_sklearn()
-#tic = time.perf_counter()
-#for n in range(nbench):
-#    qt = sklearn.preprocessing.QuantileTransformer()
-#    qt.fit_transform(X)
-#toc = time.perf_counter()
-#print(f"daal4py sklearn fit time {(toc - tic)/nbench:0.4f} seconds")
-#
-#daal4py.sklearn.unpatch_sklearn()
 
 
 afsklearn.patch_sklearn()
