+++ tests/test_sgd_classifier.py
@@ -2,16 +2,18 @@ import numpy as np
 
 from afsklearn.patcher import Patcher
 
-from . import measure_time
+from timing_utils import measure_time
 
 
 def sklearn_example() -> None:
     from sklearn.linear_model import SGDClassifier
     X = np.array([[-1, -1], [-2, -1], [1, 1], [2, 1]])
     Y = np.array([1, 1, 2, 2])
+    X = np.tile(X, (1000, 1))
+    Y = np.tile(Y, (1000))
     clf = SGDClassifier()
     clf.fit(X, Y)
-    print(f"Predict: {clf.predict([[-0.8, -1]])}")
+    print(f"Predict: {clf.predict([[-0.8, -1], [-1, -1], [-2, -1], [1, 1], [2, 1]])}")
 
 
 @measure_time
@@ -27,5 +29,7 @@ def test_afsklearn() -> None:
 
 
 if __name__ == "__main__":
+    import arrayfire as af
+    af.info()
     test_afsklearn()
     test_sklearn()
