+++ tests/test_svc.py
@@ -1,6 +1,5 @@
 # import time
 
-# import daal4py.sklearn
 # import numpy as np
 # import sklearn
 
@@ -18,20 +17,6 @@
 # toc = time.perf_counter()
 # print(f"sklearn fit time {(toc - tic)/nbench:0.4f} seconds")
 
-
-# # monkey patch scikit-learn
-# import pdb; pdb.set_trace()
-# daal4py.sklearn.patch_sklearn()
-
-# print(sklearn.svm.SVC)
-# clf = sklearn.svm.SVC(kernel='linear', C = 1.0)
-# tic = time.perf_counter()
-# for n in range(nbench):
-#     clf.fit(X,y)
-# toc = time.perf_counter()
-# print(f"intel fit time {(toc - tic)/nbench:0.4f} seconds")
-
-# daal4py.sklearn.unpatch_sklearn()
 
 # # monkey patch afsklearn
 # afsklearn.patch_sklearn()
